move AppEngineLibraries to model package and rename
